added tool-object.json
Tool-object.json refers to a specific screw available through a specific supplier. Items in the documents options block match items in the defining-tools.json options block. This follows a similar logic to that used in the Opendesk product model for product > fileset targeting. In the same way, a tool object can be matched to a hardware document when selecting specific options on a piece of hardware e.g. screw.